wip: moving compiler package
