+++ crates/lib/src/compiler/mod.rs
@@ -1,0 +1,5 @@
+//! It's the compiler!
+
+mod isa;
+pub mod quilc;
+mod rpcq;
