+++ crates/lib/src/qpu/execution.rs
@@ -17,12 +17,12 @@ use crate::qpu::{rewrite_arithmetic, runner::JobId, translation::translate};
 use crate::{ExecutionData, JobHandle};
 
 use super::client::{GrpcClientError, Qcs};
-use super::quilc::{self, CompilerOpts, TargetDevice};
 use super::rewrite_arithmetic::RewrittenProgram;
 use super::runner::{retrieve_results, submit};
 use super::translation::EncryptedTranslationResult;
 use super::QpuResultData;
 use super::{get_isa, IsaError};
+use crate::compiler::quilc::{self, CompilerOpts, TargetDevice};
 
 /// Contains all the info needed for a single run of an [`crate::Executable`] against a QPU. Can be
 /// updated with fresh parameters in order to re-run the same program against the same QPU with the
