+++ crates/lib/src/qpu/mod.rs
@@ -15,10 +15,8 @@ use tokio::time::error::Elapsed;
 
 pub mod client;
 mod execution;
-pub mod quilc;
 pub mod result_data;
 pub mod rewrite_arithmetic;
-pub(crate) mod rpcq;
 pub mod runner;
 pub mod translation;
